Comitando build 1.0.5_v.

+++ src/main/java/br/com/informe/service/PessoaService.java
@@ -1,16 +1,26 @@
 package br.com.informe.service;
 
+import br.com.informe.dto.ArquivoPessoaDTO;
+import br.com.informe.dto.InformacaoPessoaDTO;
 import br.com.informe.dto.PessoaDTO;
-import br.com.informe.entity.Pessoa;
+import br.com.informe.entity.*;
 import br.com.informe.mapper.Mapper;
+import br.com.informe.repository.ArquivoPessoaRepository;
+import br.com.informe.repository.ArquivoRepository;
+import br.com.informe.repository.InformacaoPessoaRepository;
 import br.com.informe.repository.PessoaRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.web.multipart.MultipartFile;
 
+import java.io.IOException;
+import java.util.ArrayList;
 import java.util.List;
+import java.util.Map;
 import java.util.Optional;
+import java.util.stream.Collectors;
 
 @Service
 public class PessoaService {
@@ -19,11 +29,19 @@ public class PessoaService {
     PessoaRepository repository;
 
     @Autowired
+    ArquivoPessoaRepository repositoryArquivo;
+
+    @Autowired
+    InformacaoPessoaRepository informacaoPessoaRepository;
+
+    @Autowired
     Mapper mapper;
+    @Autowired
+    private ArquivoPessoaRepository arquivoRepository;
 
     @Transactional
     public List<PessoaDTO> getlAll(){
-        return mapper.listEntityToListDTO(repository.findAll(), PessoaDTO.class);
+        return mapper.listEntityToListDTO(repository.findAllByOrderByIdDesc(), PessoaDTO.class);
     }
 
     @Transactional(propagation = Propagation.REQUIRED)
@@ -32,8 +50,12 @@ public class PessoaService {
         if (optional.isPresent()) {
             Pessoa pessoaDeletada = optional.get();
           //  pessoaDeletada.setInforme(null);
+            if ( informacaoPessoaRepository.getByInformacaoAndPessoa(null,pessoaDeletada.getId()).isEmpty()) {
+                repository.deleteById(pessoaDeletada.getId());
+            }else {
+                throw new RuntimeException("Pessoa possui vínculos no sistema");
+            }
 
-            repository.deleteById(pessoaDeletada.getId());
         }
     }
 
@@ -61,10 +83,94 @@ public class PessoaService {
 
     @Transactional
     public List<PessoaDTO> getByParametros(PessoaDTO dto) {
-        List<Pessoa> retorno = repository.getByParametros(dto.getNome());
+
+        if ( dto.getNome() != null){
+            dto.setNome(dto.getNome().toLowerCase());
+        }
+
+        if ( dto.getPai() != null){
+            dto.setPai(dto.getPai().toLowerCase());
+        }
+
+        if ( dto.getMae() != null){
+            dto.setMae(dto.getMae().toLowerCase());
+        }
+
+        if ( dto.getApelido() != null){
+            dto.setApelido(dto.getApelido().toLowerCase());
+        }
+
+        if ( dto.getCpf() != null) { //remover os pontos e traços
+            dto.setCpf(dto.getCpf().replaceAll("[\\D]", ""));
+        }
+
+        List<Pessoa> retorno = repository.getByParametros(dto.getNome() , dto.getCpf() , dto.getApelido() , dto.getMae() , dto.getPai());
 
         return mapper.listEntityToListDTO(retorno,PessoaDTO.class);
 
+    }
+
+    @Transactional
+    public List<InformacaoPessoaDTO> getVinculosPessoas(Long id) {
+        List<InformacaoPessoa> vinculos = informacaoPessoaRepository.getVinculosPessoas(id);
+        if (!vinculos.isEmpty()) {
+            return mapper.listEntityToListDTO(vinculos,InformacaoPessoaDTO.class);
+        }else{
+            return null;
+        }
+    }
+
+
+    @Transactional
+    public List<ArquivoPessoaDTO> uploadFotos(Map<String, MultipartFile> allRequestParams ,
+                                              Map<String, String> allComprimidos,
+                                              Map<String,String> tituloImagem,
+                                              Long idPessoa) {
+
+        Pessoa pessoa = Pessoa.builder().id(idPessoa).build();
+        List<ArquivoPessoaDTO> retorno = new ArrayList<>();
+        allRequestParams.forEach((key, multipartFile) ->{
+            try {
+
+
+
+                ArquivoPessoa arquivoPessoa = ArquivoPessoa.builder().arquivo(multipartFile.getBytes())
+                        .descricao(multipartFile.getOriginalFilename())
+                        .pessoaArquivo(pessoa).build();
+
+                if( allComprimidos.containsKey(key) ) {
+                    //arquivo.setArquivoComprimido(allComprimidos.get(key));
+                }
+                if( tituloImagem.containsKey(key) ) {
+                    arquivoPessoa.setTitulo(tituloImagem.get(key));
+                }
+
+                retorno.add(mapper.entityToDTO( repositoryArquivo.save(arquivoPessoa) , ArquivoPessoaDTO.class ));
+
+
+            } catch (IOException e) {
+                throw new RuntimeException(e);
+            }
+
+        });
+        return retorno;
+    }
+
+
+    @Transactional(propagation = Propagation.REQUIRED)
+    public void deletar(Long id, Long idImagem) {
+        Optional<Pessoa> optional = repository.findById(id);
+        if (optional.isPresent()) {
+            Pessoa pessoa = optional.get();
+
+            List<ArquivoPessoa> listFiltro = pessoa.getArquivos().stream().filter(arquivo -> {
+                return arquivo.getId().equals(idImagem);
+            }).collect(Collectors.toList());
+
+            listFiltro.forEach(arquivo ->{
+                arquivo.setPessoaArquivo(null);
+                arquivoRepository.deleteById(arquivo.getId()); });
+        }
     }
 
 }
